Add MicrophoneBasicViewController
